Implemented time and conventions, added test packages and test cases for current implementation.

+++ src/hk/hku/cs/c7802/base/time/TimeDiff.java
@@ -2,7 +2,7 @@ package hk.hku.cs.c7802.base.time;
 
 public class TimeDiff extends TimeSpan {
 
-	public TimeDiff(int day) {
+	public TimeDiff(long day) {
 		super(0, 0, day);
 	}
 }
